move dcgan into seperate folder
